% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R
\name{as.data.table.precautionary}
\alias{as.data.table.precautionary}
\title{Convert an object of class c('precautionary','simulations') to a data.table}
\usage{
\method{as.data.table}{precautionary}(
  x,
  keep.rownames = FALSE,
  ordinalizer = getOption("ordinalizer"),
  ...
)
}
\arguments{
\item{x}{An object of class c('precautionary','simulations')}

\item{keep.rownames}{Unused; retained for S3 generic/method consistency}

\item{ordinalizer}{If not NULL, this is a function mapping the threshold
dose ('MTDi') at which an individual experiences a binary toxicity (as
recognized by the dose-escalation design) to a named vector giving dose
thresholds for multiple grades of toxicity. The names of this vector will
be taken as designations of the toxicity grades.}

\item{...}{Additional parameters passed to the \code{ordinalizer}}
}
\description{
Convert an object of class c('precautionary','simulations') to a data.table
}
