% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three3.R
\name{Cpe3_3-class}
\alias{Cpe3_3-class}
\alias{Cpe3_3}
\title{An R6 class encapsulating precalculated CPE for 3+3 design}
\description{
An R6 class encapsulating precalculated CPE for 3+3 design

An R6 class encapsulating precalculated CPE for 3+3 design
}
\details{
TODO: Explain the hierarchy of model classes, including connections
with the executable specifications set forth in \code{exec/prolog/ccd.pl}.
}
\note{
This specializes the generic method of \code{Cpe} superclass,
removing the possibility of updating the state.
}
\examples{

## ------------------------------------------------
## Method `Cpe3_3$new`
## ------------------------------------------------

# TODO
}
\references{
\enumerate{
\item Korn EL, Midthune D, Chen TT, Rubinstein LV, Christian MC, Simon RM.
A comparison of two phase I trial designs. \emph{Stat Med}. 1994;13(18):1799-1806.
\doi{10.1002/sim.4780131802}
}
}
\section{Super class}{
\code{\link[precautionary:Cpe]{precautionary::Cpe}} -> \code{Cpe3_3}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cpe3_3$new()}}
\item \href{#method-max_dose}{\code{Cpe3_3$max_dose()}}
\item \href{#method-bU}{\code{Cpe3_3$bU()}}
\item \href{#method-J}{\code{Cpe3_3$J()}}
\item \href{#method-trace_paths}{\code{Cpe3_3$trace_paths()}}
\item \href{#method-path_matrix}{\code{Cpe3_3$path_matrix()}}
\item \href{#method-path_rx}{\code{Cpe3_3$path_rx()}}
\item \href{#method-clone}{\code{Cpe3_3$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="precautionary" data-topic="Cpe" data-id="applied">}\href{../../precautionary/html/Cpe.html#method-applied}{\code{precautionary::Cpe$applied()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="precautionary" data-topic="Cpe" data-id="path_array">}\href{../../precautionary/html/Cpe.html#method-path_array}{\code{precautionary::Cpe$path_array()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="precautionary" data-topic="Cpe" data-id="path_probs">}\href{../../precautionary/html/Cpe.html#method-path_probs}{\code{precautionary::Cpe$path_probs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="precautionary" data-topic="Cpe" data-id="report">}\href{../../precautionary/html/Cpe.html#method-report}{\code{precautionary::Cpe$report()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cpe3_3$new(D)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{D}}{number of prespecified doses}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \code{Cpe3_3} object.
}

\subsection{Returns}{
A \code{Cpe3_3} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# TODO
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-max_dose"></a>}}
\if{latex}{\out{\hypertarget{method-max_dose}{}}}
\subsection{Method \code{max_dose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cpe3_3$max_dose()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Query number of doses
}

\subsection{Returns}{
Self (invisibly), unless \code{D} is missing,
in which case the top dose, an integer, is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bU"></a>}}
\if{latex}{\out{\hypertarget{method-bU}{}}}
\subsection{Method \code{bU()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cpe3_3$bU()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the \code{b} vector and \code{U} matrix
}

\subsection{Returns}{
Named list with components \code{b} and \code{U}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-J"></a>}}
\if{latex}{\out{\hypertarget{method-J}{}}}
\subsection{Method \code{J()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cpe3_3$J()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the number \code{J} of paths
}

\subsection{Returns}{
Integer number of paths
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trace_paths"></a>}}
\if{latex}{\out{\hypertarget{method-trace_paths}{}}}
\subsection{Method \code{trace_paths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cpe3_3$trace_paths(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
No-op specialization of superclass method

Since \code{Cpe3_3} has cached paths precomputed by Prolog code,
it does not need to support this method.
}

\subsection{Returns}{
Self, invisibly
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-path_matrix}{}}}
\subsection{Method \code{path_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cpe3_3$path_matrix()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Refuse superclass method
}

\subsection{Returns}{
An error
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-path_rx"></a>}}
\if{latex}{\out{\hypertarget{method-path_rx}{}}}
\subsection{Method \code{path_rx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cpe3_3$path_rx()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Refuse superclass method
}

\subsection{Returns}{
An error
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cpe3_3$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
