% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{crmh_xo}
\alias{crmh_xo}
\title{Rust implementation of \verb{dfcrm::crmh*} integrands for w==1 case}
\usage{
crmh_xo(a, ln_x, tox, nos, s, b)
}
\arguments{
\item{a}{Numeric vector of evaluation points}

\item{ln_x}{A numeric vector of dose-wise prior log-probabilities of toxicity}

\item{tox}{A numeric vector; a dose-wise tally of toxicities}

\item{nos}{A numeric vector; a dose-wise tally of non-toxicities}

\item{s}{Scalar scale factor}

\item{b}{Order of moment to calculate (0, 1 or 2)}
}
\description{
Rust implementation of \verb{dfcrm::crmh*} integrands for w==1 case
}
