% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{exact}
\alias{exact}
\title{A wrapper function supporting exact simulation of 3+3 design}
\usage{
exact(selector_factory)
}
\arguments{
\item{selector_factory}{An object of type
\code{\link[escalation:get_three_plus_three]{three_plus_three_selector_factory}},
with \code{allow_deescalation = TRUE}.}
}
\description{
A wrapper function supporting exact simulation of 3+3 design
}
