% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity_generators.R
\docType{class}
\name{mtdi_lognormal-class}
\alias{mtdi_lognormal-class}
\alias{mtdi_lognormal}
\title{A lognormal MTDi distribution}
\description{
A lognormal MTDi distribution
}
\section{Slots}{

\describe{
\item{\code{dist}}{An object of class \code{distr6::Lognormal}}
}}

