% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R
\name{phase1_sim}
\alias{phase1_sim}
\title{Override \code{escalation:::phase1_sim} to incorporate latent toxicity tolerances}
\usage{
phase1_sim(
  selector_factory,
  true_prob_tox,
  sample_patient_arrivals = function(df) cohorts_of_n(n = 3, mean_time_delta = 1),
  previous_outcomes = "",
  next_dose = NULL,
  i_like_big_trials = FALSE,
  return_all_fits = FALSE
)
}
\arguments{
\item{selector_factory}{A \code{\link[escalation]{selector_factory}} object}

\item{true_prob_tox}{A vector of toxicity probabilities for the doses
defined in \code{selector_factory}}

\item{sample_patient_arrivals}{A function implementing an arrivals process
for trial enrollment}

\item{previous_outcomes}{This may or may not apply in applications of
package \code{precautionary}}

\item{next_dose}{Undocumented}

\item{i_like_big_trials}{I didn't choose this parameter name}

\item{return_all_fits}{Don't do this}
}
\description{
Override \code{escalation:::phase1_sim} to incorporate latent toxicity tolerances
}
