% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precautionary-package.R
\name{plan}
\alias{plan}
\alias{todo}
\title{Plan}
\description{
Plan
}
\section{Document}{

\itemize{
\item Add a proper vignette detailing and exploring exact 3+3 calculations
}
}

\section{Refactor}{

\itemize{
\item Remove the \verb{$safety} component of exact trials?
\itemize{
\item Perhaps this ought to be calculated 'on the fly' by the summary method.
\item On-the-fly calculation would postpone use of ordinalizer, in keeping with
the pattern established already for (non-exact) simulations.
}
\item Should summary(EXACT)$safety bear class 'safetytab'?
\item Example for 'as.data.table.exact'
\item What is role of G function in exact.R?
}
}

\section{Extend}{

\itemize{
\item Implement exact 3+3 variant with \code{allow_deescalation=FALSE}
\item Implement rolling 6
\item Allow an accelerated titration phase
\itemize{
\item Note that this requires access to graded toxicities at simulation time,
and therefore constitutes a substantial challenge to the generalizability
of this software design.
}
\item Index \code{sims$fits} to exact outcomes in \code{A[[D]]} where appropriate
\itemize{
\item See the \code{haystack} function in \code{exact.R}
}
}
}

\section{Robustify}{

\itemize{
\item Tests comparing results from multiple CRAN packages
}
}

