% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{print.exact}
\alias{print.exact}
\title{Specialize print method for objects of class \code{\link[escalation]{simulations}}}
\usage{
\method{print}{exact}(x, ...)
}
\arguments{
\item{x}{An object of class  c("exact","precautionary","simulations")}

\item{...}{Additional arguments; ignored}
}
\description{
Specialize print method for objects of class \code{\link[escalation]{simulations}}
}
