% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R, R/exact.R
\name{print.hyper}
\alias{print.hyper}
\title{Specialize print method defined for class \code{\link[escalation]{simulations}}}
\usage{
\method{print}{hyper}(x, ...)

\method{print}{hyper}(x, ...)
}
\arguments{
\item{x}{An object of class  c("hyper","precautionary","simulations")}

\item{...}{Additional arguments; ignored}
}
\description{
Specialize print method defined for class \code{\link[escalation]{simulations}}

Specialize print method defined for class \code{\link[escalation]{simulations}}
}
