% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R
\name{print.precautionary}
\alias{print.precautionary}
\title{Specialize print method for objects of class \code{\link[escalation]{simulations}}}
\usage{
\method{print}{precautionary}(x, ...)
}
\arguments{
\item{x}{An object of class  c("precautionary","simulations")}

\item{...}{Additional arguments; ignored}
}
\description{
Specialize print method for objects of class \code{\link[escalation]{simulations}}
}
