% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R
\name{safety_kable}
\alias{safety_kable}
\title{Output a \code{kable} for a simulation summary of class \code{safetytab}}
\usage{
safety_kable(safetytab, ...)
}
\arguments{
\item{safetytab}{An object of S3 class \code{safetytab}}

\item{...}{Additional parameters passed to \link[knitr:kable]{knitr::kable}}
}
\description{
Output a \code{kable} for a simulation summary of class \code{safetytab}
}
