% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R
\name{simulation_function.u_i}
\alias{simulation_function.u_i}
\title{Get a function that simulates dose-escalation trials using latent \code{u_i}}
\usage{
\method{simulation_function}{u_i}(selector_factory)
}
\arguments{
\item{selector_factory}{Presently, this must be a \code{tox_selector_factory};
no equivalent for \code{simulation_function.derived_dose_selector_factory}
is yet implemented.}
}
\description{
Overrides \code{\link[escalation:simulation_function]{simulation_function.tox_selector_factory}}
to return a \code{\link[=phase1_sim]{phase1_sim()}} that employs latent \code{u_i} in place of
the version native to package \CRANpkg{escalation}, which merely invokes
\code{rbinom}.
}
\details{
This function is exported for the purpose of effecting this override,
and is not meant to be invoked directly by the user.
}
