% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.R
\name{summary.exact}
\alias{summary.exact}
\title{Summarize an exact treatment of a dose-escalation design}
\usage{
\method{summary}{exact}(object, ordinalizer = getOption("ordinalizer"), ...)
}
\arguments{
\item{object}{An object of class 'exact'}

\item{ordinalizer}{An ordinalizer function}

\item{...}{Additional parameters passed to the ordinalizer}
}
\description{
Algorithmic (or 'rule-based') dose-escalation designs admit exact computation
of their outcomes. This method summarizes such an exact treatment in a manner
roughly parallel to that of \code{summary.precautionary}.
}
