% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance.R
\name{summary.precautionary}
\alias{summary.precautionary}
\title{Specialize a method defined in package 'escalation' for class 'simulations'}
\usage{
\method{summary}{precautionary}(object, ordinalizer = getOption("ordinalizer"), ...)
}
\arguments{
\item{object}{An object of class c('precautionary','simulations')}

\item{ordinalizer}{An ordinalizer function}

\item{...}{Additional parameters passed to the ordinalizer}
}
\description{
Simulations produced by package \code{precautionary} incorporate a \sQuote{u_i}
latent toxicity tolerance that characterizes the toxic dose-response of each
simulated individual trial participant. In conjunction with an
\sQuote{ordinalizer} function, these extra data enable questions to be asked
about trial safety, in terms of the probabilities of high-grade toxicities.
This function specializes the \code{escalation:::summary.simulations} method
accordingly.
}
