% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxicity_generators.R
\name{test_draw_samples}
\alias{test_draw_samples}
\title{Test performance of \code{draw_samples} function}
\usage{
test_draw_samples(n = 500)
}
\arguments{
\item{n}{Number of samples to draw}
}
\description{
Test performance of \code{draw_samples} function
}
